REFACTOR: change Login to commponent not page, useParams to Ranging
