+++ PokeDex/src/components/shared/index.js
@@ -1,5 +1,6 @@
 import ArenaBtn from "./ArenaBtn";
 import FavoriteBtn from "./FavoriteBtn";
+import Login from "./Login";
 import PokeCard from "./PokeCard";
 import PokeDetails from "./PokeDetails";
 import PokemonList from "./PokemonList";
@@ -11,6 +12,7 @@ import Wrapper from "./Wrapper";
 export {
   ArenaBtn,
   FavoriteBtn,
+  Login,
   PokeCard,
   PokeDetails,
   PokemonList,
