+++ PokeDex/src/components/subpages/Header/Header.jsx
@@ -2,8 +2,7 @@ import { useContext, useEffect } from "react";
 import { LoginContext, PokemonsListContext } from "../../../context";
 import useSyncData from "../../../hooks/useSyncData";
 import useGetPokemonData from "../../../hooks/useGetPokemonData";
-import { Wrapper } from "../../shared";
-import { Login } from "../";
+import { Login, Wrapper } from "../../shared";
 import { NavButton, Navbar } from "./Navbar";
 import Logo from "./Logo";
 import ThemeChanger from "./ThemeChanger";
