+++ PokeDex/src/components/subpages/index.js
@@ -3,7 +3,6 @@ import Edit from "./Edit/Edit";
 import Favorites from "./Favorites/Favorites";
 import Header from "./Header/Header";
 import Home from "./Home/Home";
-import Login from "./Login/Login";
 import NotLogged from "./NotLogged/NotLogged";
 import Pokemons from "./Pokemons/Pokemons";
 import Ranking from "./Ranking/Ranking";
@@ -15,7 +14,6 @@ export {
   Favorites,
   Header,
   Home,
-  Login,
   NotLogged,
   Pokemons,
   Ranking,
